Add Game API Functionality, Live Scoreboard
API integration: When a user logs in, a new game is registered with the
API. Also retrieves a history of completed games for the user and
displays the results of those games on a new scoreboard, included in the
game view sidebar. Each move updates the server with the status of the
game.

UI: Added a game reset button which displays after a game ends, which
resets the status of the board and creates a new game and sends it to
the API.